+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/Notifier.java
@@ -6,12 +6,15 @@ import com.intellij.openapi.project.Project;
 
 public class Notifier {
 
+    private static final String NOTIFICATION_GROUP = "jasypt encryption notifications";
+    private static final String JASYPT_ENCRYPT_NOTIFICATION_TITLE = "Jasypt encrypt text plugin";
+
     private Notifier() {}
 
     public static void notifyError(Project project, String content) {
         NotificationGroupManager.getInstance()
-                .getNotificationGroup("jasypt encryption error")
-                .createNotification("Jasypt encrypt text plugin", content, NotificationType.ERROR)
+                .getNotificationGroup(NOTIFICATION_GROUP)
+                .createNotification(JASYPT_ENCRYPT_NOTIFICATION_TITLE, content, NotificationType.ERROR)
                 .notify(project);
     }
 }