+++ src/test/java/org/thermoweb/intellij/plugin/encrypt/CipherUtilsTest.java
@@ -2,18 +2,24 @@ package org.thermoweb.intellij.plugin.encrypt;
 
 import org.junit.Test;
 import org.thermoweb.intellij.plugin.encrypt.cipher.CipherUtils;
-import org.thermoweb.intellij.plugin.encrypt.exceptions.JasyptPluginException;
+import org.thermoweb.intellij.plugin.encrypt.vault.CipherConfiguration;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 public class CipherUtilsTest {
 
 	@Test
-	public void should_encrypt_and_decrypt_string_correctly() throws JasyptPluginException {
+	public void should_encrypt_and_decrypt_string_correctly() {
 		String secretString = "secretString";
 		String password = "password";
-		String encryptedText = CipherUtils.encrypt(secretString, password, Algorithms.PBEWITHHMACSHA1ANDAES_128.getCode());
-		String decryptedtext = CipherUtils.decrypt(encryptedText, password, Algorithms.PBEWITHHMACSHA1ANDAES_128.getCode());
+		CipherConfiguration cipherConfiguration = new CipherConfiguration(Algorithms.PBEWITHHMACSHA1ANDAES_128, IvGenerators.RANDOMIVGENERATOR, password);
+		Result<String, CipherUtils.CipherError> encryptionResult = CipherUtils.encrypt(secretString, cipherConfiguration);
+		assertTrue(encryptionResult.isSuccess());
+		String encryptedText = ((Result.Ok<String, CipherUtils.CipherError>) encryptionResult).value();
+		Result<String, CipherUtils.CipherError> decryptionResult = CipherUtils.decrypt(encryptedText, cipherConfiguration);
+		assertTrue(decryptionResult.isSuccess());
+		String decryptedtext = ((Result.Ok<String, CipherUtils.CipherError>) decryptionResult).value();
 		assertEquals(secretString, decryptedtext);
 	}